Day 11 part 2

+++ src/main/java/org/example/day11/Item.java
@@ -1,4 +1,4 @@
 package org.example.day11;
 
-public record Item(int worryLevel) {
+public record Item(long worryLevel) {
 }
